separated database and application testing

+++ features/environment.py
@@ -44,4 +44,5 @@ def before_feature(context, feature):
 
 def after_feature(context, feature):
     os.environ['TESTING'] = '0'
-    os.remove('project/test.db')
+    if os.path.isfile('project/test.db'):
+        os.remove('project/test.db')
